Main implementation modified
The implementation of the algorithm has been modified. Note that, the algorithm has not been changed. The changes in the main function are due to the introduction of a new method in class Train. Also, the change in the interdependence of the classes is now reflected in the implementation of the main function.